Move fieldmask up to serving
